Refactor language normalization in background script and update manifest
- Removed the old language normalization logic and replaced it with a new utility function for better maintainability.
- Added a new import for language normalization in the background script.
- Updated the manifest to include the new language normalization utility script.
